Autoupdate of 2 submodules
*art updated to 80ae94ad4a4fef5341bab65c4d370e8cdb586663
Build API stubs instead of implementation jars for i18n and Conscrypt
deps.

With the thin master-art manifest we rely on the prebuilts for those
modules. They contained the implementation jars as part of the SDKs by
mistake, but that is no longer the case so those dependencies will
fail.

Test: m installclean && \
      rm -rf out/{target,host}/common/obj && \
      art/tools/buildbot-build.sh && \
      art/tools/buildbot-teardown-device.sh && \
      art/tools/buildbot-cleanup-device.sh && \
      art/tools/buildbot-setup-device.sh && \
      art/tools/buildbot-sync.sh && \
      art/tools/run-libcore-tests.sh --mode=device --variant=X64 && \
      art/tools/run-libcore-tests.sh --mode=host -j48 && \
      art/tools/run-libjdwp-tests.sh --mode=device --variant=X64 && \
      art/tools/run-libjdwp-tests.sh --mode=host && \
      art/test/testrunner/testrunner.py --target --64 && \
      art/test/testrunner/testrunner.py --host --64
Bug: 233062411
Change-Id: I8484c17e067c6f744b0715b4cdd73d9b9efa8757

*external/vogar updated to a2bdb421dc0f8b8f141e4493971f34d25b9582ba
Fix Vogar to use API stubs for dependencies from other APEX modules.

Do not depend on implementation class jars when building device tests,
which we won't have when building from prebuilts in the master-art
reduced manifest.

Test: rm -rf out/target/common/obj && \
      art/tools/buildbot-build.sh --target --skip-run-tests-build && \
      vogar --mode=device \
        libcore/ojluni/src/test/java/{lang/String,net,security}
Bug: 233062411
Change-Id: I007a767d798b1bbfefdf3ee0a72ef26346aff2a5