Read name of LE devices after connecting
LE devices can advertise name that is different from the actual name of
the device. Read name after pairing to make sure we use correct one.

This is needed for LE audio, where speaker would append location (i.e.
Left or Center) to the name in advertisement. We should then read the
exact device name to ensure it's properly displayed in settings without
the appended location.

Bug: 150670922
Test:  pair with LE device
Change-Id: I70727bd0711235929279abd1080d70deba74f502