Use genrule to pull in bindgen-erated source
Rather than having a local modification of the upstream Rust source,
instead use a genrule to replace the `${INCLUDES}` placeholder with
a re-export of everything from the bindgen-created code.

This change means that the upstream BoringSSL source no longer has a
local modification, making it easier to import new versions.

Also add a rust_test for the bindgen layout tests along the way.

Bug: 177080016
Test: build targets, atest libbssl_sys_raw_test
Change-Id: Iabf1b6ac4ae7d3a391cc038b29f157347fb36f2f