fuzzing: Add else clause to fix eager compilation
Previously, getUniform was modified to handle
uint8_t, int8_t, and bool for ChromeOS's LLVM roll
to r458507.

Apparently this slipped by the compiler failures until
actually testing the SDK on hardware. This is now
failing because the compiler will eagerly evaluate
the static_assert in uniform_int_distribution,
even though there's a return above preventing the
static_assert from being violated.

This commit adds the else to explicitly prevent
the eager static_assert evaluation.

Just like before, this is necessary but not sufficient
to verify there are no other failures in ml-nn-vts.

Bug: 231351802
Test: mma
Change-Id: Id44dd0aad13d91dd1f7405b83b1819aa3ab6722a