Refactor bcc::Chain.
This is a first step; I'm planning to do something similar to Payload.

We extract bcc::Chain to a sub-module. Then change Chain to hold the
higher level abstractions rather than the raw COSE values. Change the
public interface to perform parsing and validation in a single step,
using RawChain internally to hold the parsed values.

Push reading the file out to the client, but move Display support into
Chain.

Move unit tests next to the code they are testing.

Bug: 225177477
Test: atest ComposTestCase#testOdrefreshSpeed
Test: bcc_validator verify-chain --dump testdata/bcc/valid.chain
Test: atest libcert_request_validator_tests
Change-Id: I6c1c2b870d36274bbe4fcc249c4d8be8f5ac56ff