Remove file for reporting Fuchsia version
Remove the logic to write to a file with the Fuchsia version number.
This information is already exposed through the DUT_VERSION as part of
FlashTest. Once FlashTest is removed and the test orchestrator flashes
the device before test, this information will be attached directly to
Sponge logs at the end of testing rather than being attached during
test.

As far as I could tell, this file was not being used for anything other
than assisting during debugging. However, the introduction of
DUT_VERSION should have deprecated this file.

Fix: 234176076
Change-Id: I800791627d77c4840c856a5cf828af748acf7ae7