Consider LE Audio device in ActiveDeviceManager
When an A2DP or HFP profile is connected,
we set it as active device even when there is
an active LE audio device.
Which resulted in two active devices present.
This CL fixes it and verify it by added tests.

Tag: #stability
Bug: 233710698
Test: atest ActiveDeviceManagerTest && manually
connect BT device while BLE HS is connected.

Change-Id: I11c901da07a9ca6423942cd1702f00d557f1d0b6