Add transitive Java libraries to the runtime classpath
Java 'static' build modules can depend on non-static libraries via the Soong `libs` attribute. The latter needs to be on the test's runtime classpath since they are not present in the final jar artifact.

This change adds a new `static_deps` Bazel rule attribute and supporting logic to add all a test's transitive runtime dependencies to the runtime classpath. A new attribute is required since the outputs of these 'static' dependencies must not be added to the test's 'runfiles' seeing that they are already present in the built jar artifact.

With this change, a transitive `libs` dependency such as `junit` no longer needs to be relisted in the top-level test target since it is automatically included on the runtime classpath. Besides the usability improvements, this could also lead to smaller test jars since these commonly used artifacts are not repackaged in the final artifact.

Note that this feature is disabled by default and requires passing the `--experimental-java-runtime-dependencies` flag on the Atest command line.

Test: atest-src atest-unittests -c --host
Test: atest-src -c -m --host \
          --host-unit-test-only \
          --bazel-arg=--test_timeout=3000
Bug: 220263783
Change-Id: I1481f9e5b560419374cb209a0175e5541023ceae