Java out/inout params shouldn't be null
Added validation for out/inout parameters.

Behavioral change: without the CL, when a client calls a method with null
for out parcelable parameter, the remote side is called with non-null
object (default constructed) and the client throws NPE after
transaction. With the CL, the client will experience NPE before the
transaction.

Bug: 216045714
Test: golde_tesh.sh check
Test: aidl_integration_test
Change-Id: Iaf37955479ef2149e9a77ba1bf1fe90ca898f84f