Test registration status of device directly.
This change adds a test which allows for direct testing of whether or
not the device's public key is registered. This test is separated from
the other tests under RemoteProvisionerUnitTests and under its own build
target to avoid ambiguity. RemoteProvisionerUnitTests test that the
functionality of the device's remote provisioning implementation is
working as expected, whereas RemoteProvisionerRegistrationTest is
checking the status of the device registration on the backend server.

Since the backend server will return test certificates to an
unregistered device if the build is anything other than 'user', the test
will also edit the fingerprint to make the device appear to be running a
user build.

Test: atest RemoteProvisionerRegistrationTest
Change-Id: I1740f5f37b7c8b20812b88b3aa2f6825c5d711ff