Autoupdate of 2 submodules
*external/boringssl updated to 1710f81be4383c5456cf419abdc1712a5e25aa2e
Merge "ANDROID: Build fix for hmac.c"
*external/u-boot updated to ea3047e21e62ca51512a139d144474d6e51318a1
Bump boringssl build to gnu11

Upstream requires c11 support now, but when combining this with U-Boot's
version of the C standard library, we need to use gnu11 which provides
loff_t and uint64_t, as these things aren't strict ANSI.

Interestingly, boringssl uses "static_assert" directly, but this is a
c99-ism and was removed in c11+ and renamed _Static_assert so as to not
conflict with C++'s language feature. Just hack in a remap to keep it
happy, which is similar to what glibc does.

Bug: 234880165
Change-Id: Idce599775ce92f9bf80c6078d0e4146a1f7a216f