Create script to unzip sysroot from konan prebuilts
This CL adds a new script that can unzip the native libraries
from the KMP prebuilt. This is necessary because our build
machines have very old libraries.

It needs to be invoked by a kotlin version and out directory
pair. The script does only the extraction part, setting it to
the library path is still the caller's responsibility.

Example usage looks like:

./copy-linux-sysroot.sh --kotlin-version 1.6.21 --target-dir $OUT_DIR/syslibs
export LD_LIBRARY_PATH=$OUT_DIR/syslibs # or pass as an argument to gradle

Bug: b/228321095
Bug: b/229746360

Change-Id: I109bdeefb92f583b77ec7c701dcd4b855faf7e6d