Always get mount flags for the source of a bind mount.
mount_one was assuming that |original_mnt_flags| was always populated,
even if the destination of a bind-mount existed.

setup_mount_destination, on the other hand, exits early when the
destination exists, assuming that if the destination exists the source
flags are not needed. This is incorrect, since bind mounts are
routinely used to make a filesystem object read-only only inside a
specific mount namespace. In order to make something read-only using
a bind mount, Minijail needs to perform two mounts: one bind-mount,
and a remount with MS_RDONLY. Without knowing the flags of the original
mount, the remount cannot replicate the original mount.

Fix this by always obtaining the flags of the source, even if the
destination exists. Do this by extracting the flag-getting
functionality into a separate function. Realistically, functionality
about mount flags for the *source* has nothing to do being called in a
function referring to the destination.

Bug: 235145151
Test: Repro in bug no longer works.
Change-Id: Ia5166fdd3f36a4a01d86271e8e2dc12d6eb148ab