Revert "Generate soong_config_module_type inline in the snapshot"
This reverts commit d29f8e059712d28346eb2b92ef7b5780695ed3ad.

Although I tested this well using unit tests they don't test 
everything and although I tested the unpacking of some snapshots I did
not test what happens when all the ART sdk/module_exports
snapshots are unpacked. I just did and the build broke.

The build broke because while the soong_config_module_type modules are
marked as being Nameless (and so according to the documentation do not
have to have unique names) it turns out that is not true.

There is at least one place in the build where the names of Nameless
modules are checked and as a result this breaks.

I could fix the build breakage reasonably easily but the problem is 
that I would have to apply that fix on all partners otherwise they
wouldn't be able to receive the next set of snapshots which is too 
high a cost.

So, I am reverting this and will limit the use of inline 
soong_config_module_type modules to optional modules.

Reason for revert: Breaks the build when snapshots are unpacked.

Bug: 233965247
Change-Id: I1faae863c48f1fbfd49d2545f242401bdd4f8f8a