Use compile_multilib: "both" with "32"/"64" suffixes by default for ART native standalone tests.
It cannot be used for shared testing libraries since the suffixes
aren't picked up by DT_NEEDED or dlopen calls. They need to continue to
be installed into different directories, just like before.

This relands https://r.android.com/2119389 with a fix for
https://ci.chromium.org/ui/p/art/builders/ci/host-x86_64-cms/7684/overview
and b/235614072: Only apply the suffixes to standalone tests, to leave
test .so files and test binaries in the testing APEX alone.

Test: gtests and run tests on host and in the buildbot-build.sh chroot
Test: art/libnativebridge/tests/runtests.sh
Test: atest art_standalone_dex2oat_tests
Test: art/build/apex/runtests.sh
Bug: 233550842
Change-Id: I2b7d7327fc846b51ab91159959b2dd4bc2f503ca