Pass writeToParcel flags to writeTypedList
Parcelable.writeToParce() accepts the int flag.  But since
Parcel.writeTypedList(List<T>, int) was hidden and the other variant
passes 0 always we were not able to pass the flag according to the
calling context.

This change fixes it by passing the flag to the new API which accepts
the flag. For older SDK versions, it's done via inline helper code.

Bug: 215654054
Test: aidl_integration_test (no regression)
Change-Id: Ib5bf67195b91663dfee03cdb86f4a29579f4d18d