tp: rewrite startup metric to be much easier to understand and extend
This CL overhauls the whole startup metric to be oriented around
functions which allows much better composability than having big
monolithic views.

It also makes debugging much easier as errors in functions will be
scoped to just this function.

There should be no behavioural change as a result of this CL.

Bug: 234546650
Bug: 190219056
Change-Id: Ifc747d11d6a58199956b1cfef7d7e2c8ba9a2dc0