Package workspace resources with the Atest binary
The Atest prebuilt currently references resources from the source tree which could be incompatible with the running version. For example, adding an attribute to a Bazel rule could cause errors if the Atest code does not write that attribute in the generated target definitions.

This change resolves the situation by packaging all such resources as data dependencies of the Atest binary.

Test: atest-src atest_unittests
Test: m atest && atest-dev atest_unittests
Bug: 234519347
Change-Id: I4c3baff5be2398f4a0f6d99d1ff18cf821f20514