Autoupdate of 2 submodules
*external/linux-kselftest updated to 8dd193681b5f047f9dbd206145c9833a61d91746
futex: disable futex_wait and futex_waitv

Both tests depend on System V shared memory functions which is
disallowed on Android.  futex_waitv is implemented in 5.16 so it is
currently not available on Android as well.  Also, reverts futex_wait
and futex_waitv to upstream.

Bug: 234151152
Test: atest vts_linux_kselftest_x86_64:futex_functional_run.sh_x86_64#futex_functional_run.sh_x86_64
Test: atest vts_linux_kselftest_x86_32:futex_functional_run.sh_x86_32#futex_functional_run.sh_x86_32
Change-Id: Ia6a170bb1b69fdd5da5014d4727a0485ecbf82bb
Signed-off-by: Edward Liaw <edliaw@google.com>

*test/vts-testcase/kernel updated to 08e512b88a7262318491294c0e643327f025c01a
kselftest: futex: set color to bypass tput

tput util is not on Android, and we don't need color in the test output

Test: atest vts_linux_kselftest_x86_64:futex_functional_run.sh_x86_64#futex_functional_run.sh_x86_64
Test: atest vts_linux_kselftest_x86_32:futex_functional_run.sh_x86_32#futex_functional_run.sh_x86_32
Bug: 234151152
Signed-off-by: Edward Liaw <edliaw@google.com>
Change-Id: If46f053397dabb6d14bd81b50c15225150bf6903