Added NDK/Rust Versioning Test Coverage
V2 of IFooInterface previously introduced the function newApi().
A test case was added to the NDK and Rust suites with the expected
behavior that an instance with the V1 interface would return a
"NOT OK" status when newApi() was called on it. This was continued
from aosp/2138375. Additionally, two testcases for
acceptUnionAndReturnString() were copied over from the cpp suite
to the NDK suite.

Test: aidl_integration_test
Bug: 237455005
Change-Id: I6e2519f2789c80c8d080c28e98cc3ab770a9d6d2