Make artd tests work well enough on master-art.
Enable it for target.

The non-standalone version is enabled for host too, but it only builds
correctly in a full tree where libbinder_ndk works. To avoid fiddling
with soong_config_variables conditionals, it's still enabled on
master-art. That works as long as there's no dependency on it, since
SOONG_ALLOW_MISSING_DEPENDENCIES is used, and we only add it to the
host gtest list if we can find the binder sources.

Test: art/build/apex/runtests.sh
      art/tools/buildbot-build.sh
      m build-art-host
      m test-art-host-gtest
  on master-art and AOSP master
Test: ./art/test/testrunner/run_build_test_target.py -j80 art-gtest
      env DIST_DIR=out/dist ./art/tools/dist_linux_bionic.sh -j80 \
        com.android.art.host
  on master-art
Test: m test-art-host-gtest-art_artd_tests
  on master-art and check that the build target isn't found
Test: m test-art-host-gtest-art_artd_tests
  on AOSP master and check that the build target is found
Bug: 235464166
Change-Id: I29c5372e4d9fce468013823adc7e49da9d3332b4