Fix DW_OP_bra handling.
The handling of the DW_OP_bra is incorrect. The spec states:

DW_OP_bra is a conditional branch. Its single operand is a 2-byte
signed integer constant. This operation pops the top of stack. If
the value popped is not the constant 0, the 2-byte constant
operand is the number of bytes of the DWARF expression to skip
forward or backward from the current operation, beginning after
the 2-byte constant.

The problem is the code still used the operand even if the constant
is zero.

Modified the unit test to verify that the operand is ignored
if the stack value is zero.

Bug: 237751512

Test: Unit tests pass.
Change-Id: I012fda541b40ea65a381963c589ad7c0334e88f3