Load remote log paths from GceAvdInfo
Remote virtual devices' log paths vary with device type, multi-tenancy,
CVD tool version, etc. Acloud reports the log paths in the report file.
TradeFed reads the paths from the report, downloads them, and invokes the
loggers. If the report does not contain logs, TradeFed falls back to the
hardcoded paths.

The acloud report includes the following logs that are not in the
hardcoded paths:
- Multi-tenancy cuttlefish
- Goldfish with preconfigured IP addresses

Test: ./run_tradefed_tests.sh
Bug: 154175542
Change-Id: Ifc2d822103d7e3f87da91dd3baef010b5cdaa60f