Correct documentation for GetUsageStats
This change revises the documentation for the GetUsageStats method added
in ag/18818841. The comment previously mentioned that the method could
be used in the case that Smaps() was called; this is incorrect because
Smaps() will necessarily populate the '_usage' object.

Test: N/A, comment update only.
Bug: 235410672
Change-Id: I079ccd703269390d6233db03a4ab8aa9f85c2535
Merged-In: I079ccd703269390d6233db03a4ab8aa9f85c2535