Distinguish bind mounts requested via minijail_bind().
Minijail tries to be helpful when the user requests a bind mount that
makes a filesystem object read-only: because this requires two mounts
(one for the bind mount and one for marking RO), Minijail will perform
both mounts.

However, the minijail_mount() API can also be used for bind mounts, and
in this case Minijail also tries to be helpful, but probably
incorrectly: in many cases the callers of minijail_mount() might want
to perform the remounts themselves.

Turning one minijail_mount() call into two mount(2) system calls was
causing all sorts of trouble, and is probably breaking the principle of
least surprise.

Fix this by repurposing the unused |type| field to signal that the
bind mount was requested via minijail_bind(), and only try to perform
the necessary remount in that case.

This does temporarily re-allow bind mounting something RO as RW. Fixing
this is tracked in b/238363866.

Bug: 235145151
Test: New unit test.
Test: security.Minijail*, security.RunOCI, arc.ContainerMount pass.
Change-Id: I71b8704e8da99ad99079e995c9c84216b70789b1