tools/create_base_image_arm.sh: Start e2fsck in non-interactive mode
If non-interactive mode is not requested e2fsck's default behaviour
is to ask the user various "are you sure?" type questions before
conducting various actions. Since this script is executed in a
non-interactive environment, we receive the following error message:

"e2fsck: need terminal for interactive repairs"

To fix this, we need to call e2fsck with the flag that avoid e2fsck
asking interactive questions.

Signed-off-by: Ying-Chun Liu (PaulLiu) <paul.liu@linaro.org>
Reviewed-by: Lee Jones <lee.jones@linaro.org>
Change-Id: Ie4689269fd9e25cbe26707493797dc7860e4c748