Include mainline module origins in API levels
Update the API levels output (api-versions.xml) to include information
about which mainline module contributed what APIs. Encode this as a new
attribute:

  module="<name-of-mainline-module>"

metalava does not support top-level methods of fields. Because of this,
only add the new attribute to classes, and rely on the
inherit-if-identical semantics metalava uses for other XML attributes to
propagate the module attribute to methods and fields.

Bug: 238721263
Test: ./gradlew test --tests 'ApiGeneratorTest*Extract*System*API'
Change-Id: I5a0fb75850fb9561157765cffe08cd296a0effdc