Autoupdate of 2 submodules
*device/generic/goldfish-opengl updated to 485e4be9eac0042a85169a51a04d38112efc6571
Support AHARDWAREBUFFER_FORMAT_BLOB buffers

... by using `Buffer`s and not `ColorBuffer`s for blobs.

Bug: b/234513607
Test: launch_cvd --gpu_mode=gfxstream
Test: launch_cvd --gpu_mode=gfxstream (with WIP ANGLE)
Test: cts -m CtsNativeHardwareTestCases
Change-Id: Iab2ff38ee3e385ecd206e1d2aa6471b92985c42a

*device/generic/vulkan-cereal updated to 6402a1b33d6924677e9b6051e243119011e8f823
Support AHARDWAREBUFFER_FORMAT_BLOB buffers

by mapping BLOBs to `Buffer`s instead of `ColorBuffer`s and by
adding the needed plumbing for reads and updates.

Note: this is not a fully complete solution as technically a
guest not running with ANGLE could still have native Vulkan
users which import the BLOB AHBs. This has not yet been observed
so this might be okay for now.

Bug: b/234513607
Test: launch_cvd --gpu_mode=gfxstream (with WIP ANGLE)
Test: cts -m CtsNativeHardwareTestCases
Test: ANGLE's `ExternalBufferTestES31.*/ES3_1_Vulkan` tests
Change-Id: I782559e5a57cadc9568d46b2e0949924a6493810