Use context specific ring for QSRI export task
... as the guest assumes fence context support in
exportSyncFdForQSRILocked(). Without this, the task and fence end
up on separate rings and the fence may be signaled before the task
is completed. This can lead to unbox_VkImage() fatal failures if
the guest thinks the image is no longer in use (due to fences
being signaled early) and destroys images.

Bug: b/238778162
Test: dEQP-EGL.*
Change-Id: I7c0ec144215f4ec212b8603277114fa1c4f11b6c