Dedup *fragments information common to apex and sdk
Previously, both the sdk and apex had to specify the same *fragments
property to ensure that building a system image from a prebuilt
version of the module (both APEX and sdk snapshots) would work.

This change avoids the duplication by adding the apex to the sdk which
allows the sdk to automatically export any APIs and related information
provided by the APEX. At the moment that just includes information from
*fragments properties but may expand in future.

This also removes conscrypt.module.public.api and
conscrypt.module.platform.api from the java_sdk_libs property of the
sdk as those modules are listed in the APIs of
com.android.conscrypt-bootclasspath-fragment and so are automatically
added to the sdk when the fragment is added.

The conscrypt.module.intra.core.api module has not been removed as it
is not listed in the APIs of the fragment.

Bug: 232401814
Test: # Build snapshots with a fixed build number
      BUILD_NUMBER=fixed packages/modules/common/build/mainline_modules_sdks.sh
      # Remove api diff files as they contain file stamps of generated files so
      # differ every time they are generated.
      find out/dist/mainline-sdks -name \*txt | xargs rm
      # Save the snapshots away.
      mv out/dist/mainline-sdks before-changes
      # Apply this change.
      # Repeat the first two steps above and then run the following to verify
      # that this change had no effect on the generated snapshot contents.
      meld before-changes out/dist/mainline-sdks
Change-Id: Iddff9a621ca2cf4f1ce104f2b84ea1513417c66a