Keep all fields and methods of classes extends SystemServer
Before the change, the keep rule only keeps the class that extends
SystemServer, but if the methods/fields in the class are not used by
other entries, then the methods/fields will be removed. It will cause
issue for some service are not only used by calling the public methods
inherited from SystemServer. This change will keep all methods, and
fields in the class that extends SystemServer.

Bug: 233746531
Test: Build modules using this configuration locally, and check all
neseccary methods are kept.

Change-Id: Ia938fe8726ac8d4c1f2066f1ebca318b624e30de