Generate slim TZif chunks in the tzdata file.
That strips out 32-bit data sections. That does NOT change TZif files format within the tzdata file. I expect no issues with abseil/golang/swift.

Tests were run on CF x86_64 and x86. 32-bit data section is skipped in Java code, so Java side should not be affected by this change. On x86 mktime returns -1 for dates after 2038. That was verified even before this change in bionic tests [1].

For dates before the year 2038 there is a TimeApisConsistencyTest [2], which ensure that dates in 32-bit time_t range are handled properly.

As these tests are passing fine, I assume no APIs are broken. This is the last U tzdata change I am planning to land, so if anything happens it is fine to revert.

[1] http://cs/aosp-master/bionic/tests/time_test.cpp;l=212;rcl=ce5a9a9ffa4ff975b705f371db36e55d8056dacb

[2] http://cs/aosp-master/libcore/luni/src/test/java/libcore/java/time/TimeApisConsistencyTest.java;l=51;rcl=ce5a9a9ffa4ff975b705f371db36e55d8056dacb

Bug: 159583160

Test: CtsLibcoreTestCases
Test: CtsLibcoreOjTestCases
Test: CtsIcuTestCases
Test: CtsBionicTestCases
Test: CtsTextTestCases

Change-Id: Idcab7f1e818dd10e0bc22d7d06ca281e40cdf5a3