Autoupdate of 6 submodules
*device/generic/goldfish-opengl updated to 03ee263700531c19bd906309586ad0defc001499
Locally fork AOSP master vulkan headers

We can't touch the platform's copy of the headers because they're used
to generate libvulkan.so in the system and vendor images, and this is
NDK ABI stable. Since gfxstream needs just the headers and happily uses
them from another C path (this is done for the emulator anyway), just
fork the headers from AOSP master into the android11-gsi version.

Bug: 216827224
Change-Id: Ia3e0864b0a6e18e1d8be7d2d14e1256dc1f43d2f

*device/google/cuttlefish updated to b531a6039c3d4971f7eb0b193122d70e71fb542a
Adds ro.vendor.hwcomposer.mode sysprop and set in Gem5

... to be able to run hwcomposer in a noop mode.

Bug: 216827224
Bug: b/226154708
Test: launch Cuttlefish in `noop` mode on Crosvm and
      check that device still reaches BOOT_COMPLETE
[adelva: dropped gem5 bits, squashed with bits of pmem prop]
Change-Id: Iba17ce089b3b89e6f6c5e8eb4c2969faa9723865

Change-Id: Ibb904472e6387ccfc692d9c6493a809ba1120254

*device/google/cuttlefish_vmm updated to 2277d56ae010b2aed650859eceab6deab7091625
Update crosvm to master ToT #2

One line change to crosvm to fix the drm_virgl GPU mode.

BLOB support is not advertised because the current code in master
explicitly broke it while we wait for a virglrenderer update.

Bug: 216827224
Test: ./rebuild.sh --gce
Change-Id: Ic90bcf4a3623ddf3d679d619ad86a5cf9909d3c3

*external/crosvm updated to 993cbfe654421a0d48f7afba21e0fcdd3fed5c8d
Build fixes for wayland upgrade

crosvm.experimental is disabled for cuttlefish on this branch, but the
Android.bp files are still reachable, so we need to fix them up.

Bug: 216827224
Change-Id: I79bee55a7062cdba11640a074f0dbeea4500b569

*external/wayland updated to 9f6e9e17867fa73c6797b5e220a5326a5cd7b20a
Merge remote-tracking branch 'aosp/master' into HEAD

Newer crosvm versions require updated wayland-protocol headers, but
this forces wayland to be upgraded as well.

Bug: 216827224
Change-Id: I45e3a75786dae8db24eb77b42afe4a8c16761f9c

*external/wayland-protocols updated to 17082c0224569f631b7a4471c5afd65ed82b84b9
Revert "Stop using deprecated functionality for managing path deps"

This reverts commit 220263a3b7598abded6509bde9464c580eb27098.

This change came in as part of the merge, but it isn't compatible with
the soong implementation on android11-gsi. Revert it.

Bug: 216827224
Change-Id: I9fcb494e322396acda2872a61fc25ff6ef1c2934