Fix libnativeloader_e2e_tests to behave more like the manual tests did before https://r.android.com/2160662.
The load failures of the libraries in /product are due to missing
uses_library dependencies, which is enforced in SDK 31 and later. To
make these tests comparable with the old ones, downgrade the test apps
to SDK 30. Another reason for it is that I haven't had any luck with
adding <uses-native-library> to the app manifests.

These results are the same as the old manual tests, which has been
verified by running them with a small patch
https://r.android.com/2166263. Full output in
http://go/paste/4596862691573760.

Test: atest libnativeloader_e2e_tests:
Bug: 137356719
Change-Id: I4c8492d36fc89654ba82f58d7a560dd0254db031