Autoupdate of 2 submodules
*art updated to 6cbd91d3e47640767fc4a004cf3e858803e48135
Move PaletteReportLockContention to a new API version.

It was added in https://r.android.com/1801129, which means it went into
T and hence isn't in the first version of the Palette API.

Also add some comments to separate the API versions, and align
libartpalette.map.txt with the one in system/libartpalette after
https://r.android.com/2164906.

Test: m
Bug: 196334695
Change-Id: I68579fe8a64022d8ee7c828609240e4e3c29b32d

*system/libartpalette updated to 87bd9b372e67de190464a6698c5161d559e390aa
Add receiver for the PaletteReportLockContention callback.

This was added to ART in T in https://r.android.com/1801129, but never
added here. Without this function the calls from ART end up in the
PaletteMethodNotSupported fallback in
art/libartpalette/apex/palette.cc, and since ART ignores the return
value there's no practical difference.

This is done to straighten out the API history and avoid differences
wrt art/libartpalette.

Bug: 196334695
Test: m
Change-Id: Ia156fc11be12b832413d7d7490a8856fb34778f9