Add a new test for Startup Trace
Validates that Startup Trace works fine if we dont emit any event
during startup tracing session.

Note: To make it work, we had to remove `PERFETTO_CHECK(!fully_bound_)`
from `SharedMemoryArbiterImpl::BindStartupTargetBuffer`.

The reason being: when we start a startup tracing session with a data
source and never trace packets from that data source, the trace writter
will never be created. The data source will be assigned a
reservation_target_buffer_id but it doesn't mean anything unless a
trace-writter is created for this data source. As a result
shared_memory_arbiter will continue to think its fully_bound, because it
was not present in its `target_buffer_reservations_` map.

While aborting such a startup trace session or adopting it to normal
session we would like to call `BindStartupTargetBuffer` for this data
source. Which is almost like assigning a target_buffer to a fresh data
source. Things work well, but we cannot let `BindStartupTargetBuffer`
assume that this method will be called only when we already have
something unbound to be bounded. We should be allowed to use this
method for `reservation_id` that shared_memory_arbiter have never seen
before (i.e. `CreateStartupTraceWriter` was never called on it). It
should just insert those newly discovered `reservation_id` in
`target_buffer_reservations_` map and do the things as it would do if
`reservation_id` was already there in this map.
This part is already implemented. It was just this CHECK(!fully_bound_)
which was blocking this use case. Hence this check is removed.

Change-Id: I00af55b73dfa3a32477d0d8a9bd78e217d5320d3