Take std::string_views instead of const std::string & for android::base::Trim
This avoids implicit conversions involving copies from non-std::string
types to std::string for argument s. The implementation is also greatly
simplified. There is still a copy for the return value as before. Retain
the old symbol to keep binary compatability, but implement it in terms
of the new function. Use a function template to retain source
compatibility for types which are implicitly convertible to std::string
but not std::string_view. Conversions which are only possible explicitly
are prohibited.

Bug: N/A
Test: atest libbase_test
Change-Id: Ib6c312ef197877a2bfc92fc946e1ba8794f55df1