ANDROID: fuse-bpf: Remove BPF fd from children inode as a part of BPF execution
According to the current logic, only the container of package-private
folders (obb/ and data/) need to have a BPF program and it is
successfully handled in the fuse daemon. But when we stay only with
the backing implementation (currently, only one possible flow) and
don’t have userspace lookups at all, all children folders
inherited BPF programs from their parents.

So handling this case and removing BPF programs from all children as
a part of postfilter.

Bug: 219958836
Test: Manually checked that we’re not receiving bpf calls on (data|obb)
Signed-off-by: Dmitrii Merkurev <dimorinny@google.com>
Change-Id: I80d28efe550c92b78a44683206cc9635fc1b8863