[protozero] Refactor nested enum values
Currently nested enums are located in the common namespace
and therefore the names of their values have a form of
OuterClass_Enum_ENUM_VALUE, which is not pleasant to use directly.

This is counteracted by an OuterClass having ENUM_VALUE field, but
it's still a problem for the folks who want to use this enum directly.

This patch addressed this by putting the enums into their own namespace
and generating compat layer:

namespace perfetto_pbzero_enum_OuterClass_Enum {
enum Enum {
   ENUM_VALUE = 0,
   ...
}
}  // namespace

R=rsavitski@google.com,primiano@google.com
Change-Id: Id33ee85544c10f99fd61fbd187e0436974df8e22