Update showmap to use maps by default
This change adds collect_vma_merge_names, which sums VMA sizes with a
map lookup instead of a linear search. This method is used by default;
the original collect_vma function is now only used when showmap is given
the -a or -v options.

This results in a ~9% reduction in cpu cycles spent when calling
showmap.

This change also fixes the method used to name unnamed VMAs by keeping
track of the most recently-inspected VMA in 'recent'.

Bug: 215574756
Test: - atest libmeminfo_test
      - Manually verified that the output of showmap aligned with the
	output of showmap -a/-v.
Merged-In: I212f058309e8c0044811ca48897fbca959403f6a
Change-Id: I212f058309e8c0044811ca48897fbca959403f6a