Update ProcMemInfo to report usage in KB
This change modifies ReadVmaStats() to report usage in KB instead of B.
This is intended to keep memory units consistent between Smaps() and
Maps(). Existing tools that call methods that use Maps() (procrank,
librank, procmem, wsstop) have been updated to reflect this.

A default KB-to-B conversion has been added to FillInVmaStats() (which
calls ReadVmaStats()), because most callers of this function expect
usage to be reported in bytes.

Bug: 236427386
Test: - atest libmeminfo_test
      - manually checked before/after output of the modified tools.
Change-Id: I036082d0e2d7a3869212e9ef0decf6ab030a69fa