Update the documentation of some `OdrefreshReported` fields.
Now that the boot image ("core image") is no longer included in
the ART APEX, `odrefresh` does recompile the entire boot
classpath (boot image + boot image extensions) -- not just the
boot image extensions. Reflect this in the documentation of
`android.os.statsd.OdrefreshReported`.

Also make it clear that the "primary" and "secondary" qualifiers
used for the boot classpath refer to the architectures supported
by the device, not the boot image and its extension(s), which
were sometimes called (resp.) "primary boot image" and "secondary
boot image".

#art-doc

Test: n/a (comment-only change)
Bug: 193773739
Change-Id: I3efff79e0866eb361f80ecebab82ee2a07a21d68