libhidl: check ro.debuggable less
We were always checking this when retrieving services, but
we don't need to (we can check it only when treble testing
override is on).

There are a few things we could consider in this codebase:
- checking debuggability when the testing override is initially
  set (and warning then) VS now when it is read
- avoiding duplicated calls to isDebuggable in
  isTrebleTestingOverride and getRawServiceInternal.
- removing 'kEnforceVintfManifest' since it is always true now.

However, given that we're deprecating HIDL and in the interest
of keeping changes simple, focusing on other things instead
of these improvements.

Bug: 193912100
Test: boot
Change-Id: I6df328dead00c05dcbd8eedf63fc24a8d6141871