Webview: Inherit IO from the webview installer tool
This CL makes the app launch test inherit all the IO
from the WebView installer tool process. All the tool's
IO will be outputted to the terminal. This replaces
the current process where we do a blocking read on the
output from the tool. This is problematic because if
the tool hangs then the app launch test wll also hang
because it is waiting for output from the tool.

Test: csuite-tradefed run webview-app-launch <params>
Bug: 206871132
Change-Id: I185e9f6d4590cee48d01abadea1880a15ec51e7a