Fix shared/static library paths
The build key contains the word static/shared depending on which type
of library it built. Hence if we use it for both the static and shared
libraries paths that static libraries will be in a folder called
..._shared_... and vice versa.

We should replace the word static/shared depending on which type of library it is.

Test: Ran samples
Bug: 241066533
Change-Id: I651c9210bda4b66ff6ea122ec6a8b473ab34cb74