vulkan-cereal: add virtgpu_gfxstream_protocol.h
This adds a centralized place for the gfxstream virtgpu
protocol.  Right now, it exists but is spread out through
random files in vulkan-cereal/goldfish-opengl.

This gives an opportunity to simplify and even modify the
protocol.  For example,  some commands have never used by
the guest (such as kVirtioGpuNativeSyncCreateImportFd +
kVirtioGpuNativeSyncVulkanCreateImportFd) and probably
won't without modifications.

Long term, it may be possible to use gfxstream-protocols +
FlatBuffers to house this.  For rapid prototyping and
modification (especially for ASG), just do this for now.

BUG=235485545
TEST=compile

Change-Id: I36cc3178dc4fe8c98f4ad87a4f78326ff2b03288